// projects/cloud-airlines-angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarOffer.ts
/**
 * Airline API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CfarItinerary } from './cfarItinerary';
import { MapString } from './mapString';
import { RequestType } from './requestType';

/**
 * A successful CFAR offer response
 */
export interface CfarOffer { 
    /**
     * Unique identifier for an offer
     */
    id: string;
    /**
     * Amount per passenger to be paid by user for CFAR
     */
    premium: string;
    /**
     * Amount per passenger to be refunded to user upon CFAR exercise
     */
    coverage: string;
    /**
     * Currency of offer
     */
    currency: string;
    requestType: RequestType;
    /**
     * USD Exchange rate for currency; amount * toUsdExchangeRate == USD
     */
    toUsdExchangeRate: string;
    /**
     * A UTC [RFC3339](https://xml2rfc.tools.ietf.org/public/rfc/html/rfc3339.html#anchor14) datetime; the date and time at which the CFAR contract will expire once purchased
     */
    contractExpiryDateTime: Date;
    /**
     * A UTC [RFC3339](https://xml2rfc.tools.ietf.org/public/rfc/html/rfc3339.html#anchor14) datetime; the date and time at which a CFAR offer was created
     */
    createdDateTime: Date;
    itinerary: CfarItinerary;
    offerDescription: Array<string>;
    extAttributes: MapString;
}